Extract :scissors: UploadBody from UploadRequest
